parts-finder/backend/requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
flask-limiter>=3.3.0
flask-session>=0.4.0
python-dotenv>=0.19.0
pyinstaller>=5.0.0
werkzeug>=2.0.0
requests>=2.28.0
pytest>=7.0.0
pytest-flask>=1.2.0
cryptography>=38.0.0